// src/components/picker/PickerItem.d.ts
import React from 'react';
import { PickerItemProps } from './types';
/**
 * @description: Picker.Item, for configuring the Picker's selectable options
 * @example: https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/PickerScreen.tsx
 */
declare const PickerItem: {
    (props: PickerItemProps): React.JSX.Element;
    displayName: string;
};
export default PickerItem;